+++ amfiteatr_core/src/util/tboard.rs
@@ -5,12 +5,12 @@ use crate::error::AmfiteatrError;
 
 pub trait TensorboardSupport<S: Scheme>{
 
-    fn add_tboard_directory<P: AsRef<std::path::Path>>(&mut self, directory_path: P) -> Result<(), AmfiteatrError<S>>;
+    fn add_tboard_directory(&mut self, directory_path: &Path) -> Result<(), AmfiteatrError<S>>;
     fn t_write_scalar(&mut self, index: i64, tag: &str, value: f32) -> Result<bool, AmfiteatrError<S>>;
 }
 
 impl<S: Scheme, T: TensorboardSupport<S>> TensorboardSupport<S> for Arc<Mutex<T>>{
-    fn add_tboard_directory<P: AsRef<Path>>(&mut self, directory_path: P) -> Result<(), AmfiteatrError<S>> {
+    fn add_tboard_directory(&mut self, directory_path: &Path) -> Result<(), AmfiteatrError<S>> {
         match self.as_ref().lock(){
             Ok(mut guard) => guard.add_tboard_directory(directory_path),
             Err(e) => Err(AmfiteatrError::Lock { description: "Add tboard directory".to_string(), object: format!("{e}") })
@@ -27,7 +27,7 @@ impl<S: Scheme, T: TensorboardSupport<S>> TensorboardSupport<S> for Arc<Mutex<T>
 }
 
 impl<S: Scheme, T: TensorboardSupport<S>> TensorboardSupport<S> for Mutex<T>{
-    fn add_tboard_directory<P: AsRef<Path>>(&mut self, directory_path: P) -> Result<(), AmfiteatrError<S>> {
+    fn add_tboard_directory(&mut self, directory_path: &Path) -> Result<(), AmfiteatrError<S>> {
         match self.lock(){
             Ok(mut guard) => guard.add_tboard_directory(directory_path),
             Err(e) => Err(AmfiteatrError::Lock { description: "Add tboard directory".to_string(), object: format!("{e}") })
@@ -44,7 +44,7 @@ impl<S: Scheme, T: TensorboardSupport<S>> TensorboardSupport<S> for Mutex<T>{
 }
 
 impl<S: Scheme, T: TensorboardSupport<S>> TensorboardSupport<S> for RwLock<T>{
-    fn add_tboard_directory<P: AsRef<Path>>(&mut self, directory_path: P) -> Result<(), AmfiteatrError<S>> {
+    fn add_tboard_directory(&mut self, directory_path: &Path) -> Result<(), AmfiteatrError<S>> {
         match self.write(){
             Ok(mut guard) => guard.add_tboard_directory(directory_path),
             Err(e) => Err(AmfiteatrError::Lock { description: "Add tboard directory".to_string(), object: format!("{e}") })